src/third_party/blink/web_tests/editing/selection/mouse/drag_over_text_transform.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#target { text-transform: uppercase; }
</style>

<p id="target">&#xDF;abc</p>

<script>
// Drag through text with 'text-transform'. Shouldn't crash.

function dragThrough(x1, x2, y) {
  return new Promise((resolve, reject) => {
    assert_own_property(window, 'chrome');
    assert_own_property(window.chrome, 'gpuBenchmarking');

    chrome.gpuBenchmarking.pointerActionSequence([{
      source: 'mouse',
      actions: [
        {name: 'pointerDown', x: x1, y: y},
        {name: 'pointerMove', x: x2, y:y},
        {name: 'pointerUp'}
      ]}], resolve);
  });
}

promise_test(async () => {
  const target = document.getElementById('target');
  const x1 = target.offsetLeft;
  const x2 = x1 + target.offsetWidth;
  const y = target.offsetTop + target.offsetHeight;

  await dragThrough(x1, x2, y);

  const selection = getSelection();
  const text = target.firstChild;
  assert_false(selection.isCollapsed);
  assert_equals(selection.anchorNode, text);
  assert_equals(selection.anchorOffset, 0);
  assert_equals(selection.focusNode, text);
  assert_equals(selection.focusOffset, 4);
});

</script>
